pyproject.toml
[project]
name = "mcp_doc_retriever"
version = "0.1.0"
requires-python = ">=3.11" 
dependencies = [
    "beautifulsoup4>=4.13.3",
    "fastapi>=0.115.12",
    "lxml>=5.3.2",
    "playwright>=1.51.0",
    "pydantic>=2.11.2",
    "requests>=2.32.3",
    "uvicorn[standard]>=0.34.0",
    "httpx>=0.27.0",
    "aiofiles>=24.1.0",
    "packaging>=24.2",
    "filelock>=3.18.0",
    "pip>=25.0.1",
    "bleach>=6.2.0",
    "sse-starlette>=2.2.1",
    "tqdm>=4.67.1",
    "pathvalidate>=3.2.3",
    "loguru>=0.7.3",
    "markdown-it-py>=3.0.0",
    "tree-sitter-language-pack>=0.7.1",
    "typer>=0.12.5", # Added for CLI functionality
    "aiosqlite>=0.21.0",
    "python-arango==8.1.6",
    "litellm>=1.66.1",
    "redis>=5.2.1",
    "json-repair>=0.41.1",
    "rapidfuzz>=3.13.0",
    "tiktoken>=0.9.0",
    "tree-sitter>=0.24.0",
    "tree-sitter-languages>=1.10.2",
    "docutils>=0.21.2",
    "sphinx>=8.2.3",
    "sphinx-rtd-theme>=3.0.2",
    "pandoc>=2.4",
    "pyperclip>=1.9.0",
    "deepmerge>=2.0",
    "async-timeout>=5.0.1",
    "tenacity>=9.1.2",
]
    

[project.scripts]
mcp-doc-retriever = "mcp_doc_retriever.cli:app"
lessons-cli = "mcp_doc_retriever.lessons_cli:app"

[project.optional-dependencies]
dev = ["ruff"]
test = [
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "httpx>=0.27.0",  
    "respx>=0.22.0",  
    "docker>=7.1.0",  
    "requests",      
    "pytest-dependency", 
    "python-dateutil", 
    "python-dotenv",   
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
markers = [
    "dependency: mark test to run after specified dependencies",
]
